15-04

+++ src/main/java/br/com/macedo/sistemas/repository/PjRepository.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import br.com.macedo.sistemas.domain.PessoaJuridica;
 
-public interface PjRepository extends JpaRepository<PessoaJuridica, Long>{
+public interface PjRepository extends JpaRepository<PessoaJuridica, Integer>{
 	
 	@Transactional(readOnly = true)
 	PessoaJuridica findByCnpj(String cnpj);
